Constants: More Consistency
The type field is the unique identifier of an error, and we’re supposed to have documentation for each type. So it’s really important to keep track of these and never misspell them.

That sounds like a perfect use-case for constants! Add a constant on ApiProblem for the validation_error key. Now, just reference the constant when instantiating ApiProblem.

Awesomely enough that’s one less spot for me to screw up.

Mapping title to type
But we can go further. According to the spec, the title field is the description of a given type. In other words, we should have the exact same title everywhere that we use the validation_error type.

To force this consistency, create an array map on ApiProblem from type to its human-description.
Note
You can also choose to translate the title. If you need this, you’ll need to run the key through your translator before returning it.

And instead of passing the  as the third argument to the constructor, we can just look it up by the . And like the good programmers we are, we’ll throw a huge, ugly and descriptive exception if we don’t find a title.
Back in the controller, we can now safely remove the last argument when constructing the ApiProblem object.
Bam! We have an ApiProblem class to keep things consistent, a constant for the one problem type we have so far, and a title that’s automatically chosen from the type.