Error in Invalid JSON Beyond validation errors, what else could go wrong? Well what if a tricky client tries to mess with us by sending invalid JSON? Right now, that would result in a cryptic 500 error message. We can’t let them catch us off guard, so we need a 400 status code with a clear explanation to tell them that we’re always watching.
Let’s write a test! I’ll copy the validation error scenario, but remove a quote so that the JSON is invalid.
For now, let’s just continue to check that the status code is 400. Take that shifty client! If we run the test immediately, it fails with a 200 error instead.