Handling Invalid JSON
In our controller, we’re already checking to see if the JSON is invalid.
To make this a 400 error, we could do 2 things. First, we could create a new Response object and set its status code to 400. That’s what we’re already doing with the validation error.

Second, in Silex, we could throw a special HttpException.

In most frameworks, if you throw an exception, it results in a 500 status code. That’s true in Silex too, unless you throw this very special type of exception where the status code is the first argument.

First, make sure it’s working by running the test:

Tip
Silex/Symfony has a collection of exception classes(https://github.com/symfony/symfony/tree/master/src/Symfony/Component/HttpKernel/Exception), one for each of the most common 400 and 500-level responses. For example, BadRequestHttpException is a sub-class of HttpException that sets the status code to 400. The result is the same: throwing these “named” exception classes is just a bonus to give your code more consistency and clarity.
Awesome! So why am I throwing an exception instead of just returning a normal 400 response? The problem is that we’re inside handleRequest, so if I return a Response object here, it won’t actually be sent back to the user unless we also return that value from newAction and updateAction. That just gets confusing and a bit ugly.

Instead, if we throw an exception, the normal execution will stop immediately and the user will definitely get the 400 response. So being able to throw an exception like this makes my code easier to write and understand. Double threat!

The disadvantage is complexity. When I throw an exception, I need to have some other magic layer that is able to convert it into a proper response. In Silex, that magic layer is smart enough to see my HttpException and create a response with a 400 status code instead of 500.

If this doesn’t make sense yet, keep following along.