ApiProblem for Invalid JSON
Since invalid JSON is a 'problem', we should really send back an application/problem+json response. Let’s first update the test to look for this Content-Type header and then look for a type field that’s equal to invalid_body_format:

# features/api/programmer.feature # ...

Scenario: Error response on invalid JSON
# the rest of the scenario # ... And the 'Content-Type' header should be 'application/problem+json' And the 'type' property should equal 'invalid_body_format'
To make this work, we’ll create a new ApiProblem object. But first, let’s add the new invalid_body_format type as a constant to the class and give it a title.
Next, instantiate the new ApiProblem in the controller.

But now what? When we had validation errors, we just created a new JsonResponse, passed {dollar_sign}problem->toArray() as data, and returned it. But here, we want to throw an exception instead so that the normal flow stops.

We’re going to fix this in two steps. First, we will throw an Exception, but we’ll put the ApiProblem inside of it. Second, we’ll hook into the magic layer that handles exceptions and extend it so that it transforms the exception into a Response with a 400 status code. Again, this is a little more complicated, so if it doesn’t make sense yet, watch our implementation.