ApiProblemException and Exception Handling In order to be able to throw an exception that results in a JSON response, we need to hit the gym and first create a new class called ApiProblemException. Make it extend that special HttpException class.
The purpose of this class is to act like a normal exception, but also to hold the ApiProblem inside of it. To do this, add an apiProblem property and override the __construct method so that an ApiProblem object is the first argument.

The exception still needs a message and I’m calling getTitle() on the ApiProblem object to get it. Open up the ApiProblem class and add this getTitle() function so we can access it.
Finally, go back to ApiProblemException and add a getApiProblem getter function. Hang tight, we’ll use this in a few minutes.