ApiProblemException for Validation
This is really powerful. If we need to return a 'problem' anywhere in our API, we only need to create an ApiProblem object and throw an ApiProblemException.

Let’s take advantage of this for our validation errors. Find handleValidationResponse and throw a new ApiProblemException instead of creating and returning a JsonResponse object. And to keep things clear, let’s also rename this function to throwApiProblemValidationException.
Now, update newAction and updateAction to use the new function name. We can also remove the return statements from each: we don’t need that anymore.
And when we run the tests, all green! Piece by piece, we’re making our code more consistent so that we can guarantee that our API is consistent.