Handling 404 Errors We’re handling validation errors and invalid JSON errors. The last big thing is to properly handle 404 errors. In showAction and updateAction, we’re throwing a special type of exception class to trigger a 404 response. But in reality, the 404 response isn’t JSON: it’s a big HTML page. You can see this by browsing to a made-up programmer:
  http://localhost:8000/api/programmers/bumblebee
And actually, if we go to a completely made-up URL, we also see this same HTML page:

  http://localhost:8000/api/foo/bar
Internally, Silex throws that same exception to cause this 404 page.

Somehow, we need to be able to return JSON for all exceptions and while we are at it we should use the API problem detail format.