Exposing more Error Details In our app, we trigger 404 pages by calling the throw404 shortcut function in a controller. Behind the scenes, this throws a NotFoundHttpException, which extends that very special HttpException class. The result is a response with a 404 status code. That magic is part of Silex. But without this, we could have built in our own logic in the exception listener to map certain exception classes to specific status codes.
All exceptions have an optional message, which we can populate by passing an argument to throw404. We’re already including some details that might help the API client.

If you go to /api/programmers/bumblebee in the browser, we get a 404 JSON response, but we don’t see that message. Look back at our exception-handling function. The NotFoundHttpException is not an instance of ApiProblemException, so we fall into the situation where we create the ApiProblem by hand. We’re not using the exception’s message anywhere, so it makes sense we don’t see it:

// src/KnpU/CodeBattle/Application.php

// ...
} else {
    {dollar_sign}apiProblem = new ApiProblem({dollar_sign}statusCode);
}
Before we fix this, let’s update the 404 scenario. We can use the standard details field to store the exception message.