Let’s throw a big exception from inside the showAction controller method and see how the tests look.
Let’s run just one of the scenarios that uses showAction. To do this, we can point Behat directly at the .feature file and include the line number where the word Scenario: appears:

  php vendor/bin/behat features/api/programmer.feature:64

When the test fails, it prints out what the client will see, which is our API Problem media type response... but with absolutely no details beyond the 500 status code.

While developing, that’s not helpful. Instead, for 500 errors, I want to continue seeing the big beautiful, normal error page, because it includes the exception message and stacktrace.